requirements.txt
torch
torchvision
fastapi
pillow
streamlit
matplotlib
